test: BlackJackResult 테스트 추가

+++ src/main/java/blackjack/domain/result/BlackJackResult.java
@@ -45,8 +45,8 @@ public enum BlackJackResult {
 		return (int)(bet * profit);
 	}
 
-	public double getReverseEarning(double earning) {
-		return LOSE.getProfit() * earning;
+	public int getReverseEarning(int earning) {
+		return (int)(LOSE.getProfit() * earning);
 	}
 
 	public double getProfit() {
